+++ src/flickr_api/__init__.py
@@ -13,7 +13,7 @@ from .exceptions import (
 )
 
 
-__version__ = "3.11.3"
+__version__ = "3.12.0"
 
 
 __all__ = [